Chunky rewrite to the CPAL interface (won't compile on Linux yet!)
