+++ audio/src/audio.rs
@@ -1,6 +1,14 @@
 use anyhow::Result;
 use symphonia::core::audio::SignalSpec;
 
+pub trait OpenOutputStream {
+    fn open(spec: AudioSpecification) -> Result<Box<dyn AudioOutput>>;
+}
+
+pub trait OpenInputStream {
+    fn open(spec: AudioSpecification) -> Result<Box<dyn AudioInput>>;
+}
+
 pub trait AudioOutput {
     fn write(&mut self, samples: &[f32]) -> Result<()>;
     fn flush(&mut self);
@@ -11,14 +19,10 @@ pub trait AudioInput {
     fn flush(&mut self);
 }
 
-pub trait AudioConfiguration {
-    fn get_outputs(&mut self) -> Vec<String>;
-    fn get_inputs(&mut self) -> Vec<String>;
-}
-
-#[cfg(target_os = "linux")]
-pub fn get_configuration() -> Box<dyn AudioConfiguration> {
-    Box::new(crate::pulse::configuration::get_configuration())
+pub struct AudioSpecification {
+    pub device: Option<String>,
+    pub spec: SignalSpec,
+    pub buffer: usize,
 }
 
 #[cfg(target_os = "linux")]
@@ -36,19 +40,11 @@ pub(crate) fn get_input(device: Option<String>) -> Result<Box<dyn AudioInput>> {
 }
 
 #[cfg(not(target_os = "linux"))]
-pub fn get_configuration() -> Box<dyn AudioConfiguration> {
-    Box::new(crate::cpal::configuration::get_configuration())
+pub(crate) fn get_output(spec: AudioSpecification) -> Result<Box<dyn AudioOutput>> {
+    crate::cpal::cpal_playback::CpalPlayback::open(spec)
 }
 
 #[cfg(not(target_os = "linux"))]
-pub(crate) fn get_output(
-    signal_spec: SignalSpec,
-    device: Option<String>,
-) -> Result<Box<dyn AudioOutput>> {
-    crate::cpal::playback::CpalAudioOutput::open(signal_spec, device)
-}
-
-#[cfg(not(target_os = "linux"))]
-pub(crate) fn get_input(device: Option<String>) -> Result<Box<dyn AudioInput>> {
-    crate::cpal::record::CpalAudioInput::open(device)
+pub(crate) fn get_input(spec: AudioSpecification) -> Result<Box<dyn AudioInput>> {
+    crate::cpal::cpal_record::CpalRecord::open(spec)
 }
