--- audio/src/cpal/record.rs
@@ -1,152 +1,0 @@
-use anyhow::{anyhow, bail, Result};
-use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
-use log::{debug, error};
-use rb::{Consumer, RbConsumer, RbProducer, SpscRb, RB};
-use std::rc::Rc;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
-use std::time::Duration;
-
-use crate::audio::AudioInput;
-pub struct CpalAudioInput;
-
-// Set max 'interim' buffer size to 200ms. This defines the maximum time between the audio being
-// handled here, and the external reader pulling it out before samples are lost.
-//
-// This may seem a little high, but the buffered recorder needs to handle thread locks, mutexes,
-// and redistributing the samples out to anything that needs them. This can lead to small delays
-// when reading. I've seen this reach 6-7,000 samples before, so this buffer should be clear.
-const BUFFER_SIZE: usize = 19200;
-
-impl CpalAudioInput {
-    pub fn open(device: Option<String>) -> Result<Box<dyn AudioInput>> {
-        // Device handling is a little more awkward on CPAL, we need the audio host and device.
-        // We request them as 'HOST*DEVICE' on the Command Line, now we can look them up :p
-
-        let mut cpal_device = None;
-
-        // Basically, if *ANYTHING* goes wrong here, we'll fall through to default.
-        if let Some(device_name) = device {
-            if let Some(position) = device_name.find('*') {
-                let str_host = &device_name[0..position];
-                let str_device = &device_name[position + 1..device_name.len()];
-
-                // Ok, now for cpal, find the correct host..
-                let cpal_host_list = cpal::available_hosts();
-                let host_id = cpal_host_list.iter().find(|x| x.name() == str_host);
-
-                if let Some(host_id) = host_id {
-                    if let Ok(host) = cpal::host_from_id(*host_id) {
-                        // We have found our host, now try to find the device..
-                        if let Ok(mut devices) = host.input_devices() {
-                            if let Some(device) = devices.find(|x| {
-                                x.name().unwrap_or_else(|_| "UNKNOWN".to_string()) == str_device
-                            }) {
-                                cpal_device = Some(device);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        let final_device;
-        if let Some(device) = cpal_device {
-            final_device = device;
-        } else {
-            let host = cpal::default_host();
-            final_device = match host.default_input_device() {
-                Some(device) => device,
-                None => return Err(anyhow!("Unable to find Default Device")),
-            };
-        }
-
-        // Select proper recording routine based on sample format.
-        CpalAudioInputImpl::try_open(&final_device)
-    }
-}
-
-struct CpalAudioInputImpl {
-    ring_buf_consumer: Consumer<f32>,
-    stream: cpal::Stream,
-    stream_closed: Arc<AtomicBool>,
-
-    // Use a shared read buffer..
-    read_buffer: [f32; BUFFER_SIZE],
-}
-
-impl CpalAudioInputImpl {
-    pub fn try_open(device: &cpal::Device) -> Result<Box<dyn AudioInput>> {
-        // Input
-        let config = cpal::StreamConfig {
-            channels: 2 as cpal::ChannelCount,
-            sample_rate: cpal::SampleRate(48000),
-
-            // Set the main read buffer at 10ms
-            buffer_size: cpal::BufferSize::Fixed(960),
-        };
-
-        // Prepare the Read Buffer, grab the producer and consumer..
-        let ring_buf = SpscRb::<f32>::new(BUFFER_SIZE);
-        let (ring_buf_producer, ring_buf_consumer) = (ring_buf.producer(), ring_buf.consumer());
-        let stream_closed = Arc::new(AtomicBool::new(false));
-        let stream_closed_inner = stream_closed.clone();
-
-        let stream_result = device.build_input_stream(
-            &config,
-            move |data: &[f32], _: &cpal::InputCallbackInfo| {
-                // Write out as many samples as possible from the ring buffer to the audio output.
-                if let Err(samples) = ring_buf_producer.write(data) {
-                    debug!("{}", samples);
-                }
-            },
-            move |_| {
-                // Instruct anything that attempts to read that this stream is gone.
-                stream_closed_inner.store(true, Ordering::Relaxed);
-                error!("Audio Input Error from CPAL")
-            },
-        );
-
-        if stream_result.is_err() {
-            return Err(anyhow!("Unable to open Stream"));
-        }
-
-        let stream = stream_result.unwrap();
-
-        // Start the input stream.
-        if stream.play().is_err() {
-            return Err(anyhow!("Unable to begin Playback"));
-        }
-
-        Ok(Box::new(CpalAudioInputImpl {
-            read_buffer: [0.0; BUFFER_SIZE],
-            ring_buf_consumer,
-            stream,
-            stream_closed,
-        }))
-    }
-}
-
-impl AudioInput for CpalAudioInputImpl {
-    fn read(&mut self) -> Result<Vec<f32>> {
-        if self.stream_closed.load(Ordering::Relaxed) {
-            bail!("Audio Stream has been closed.");
-        }
-
-        // We've put a timeout on this, because if CPAL suddenly stops sending packets due to an
-        // error, a read can block forever. 250ms should be sufficient!
-        let read = self
-            .ring_buf_consumer
-            .read_blocking_timeout(&mut self.read_buffer, Duration::from_millis(250));
-
-        if let Ok(Some(samples)) = read {
-            return Ok(Vec::from(&self.read_buffer[0..samples]));
-        };
-
-        Ok(vec![])
-    }
-
-    fn flush(&mut self) {
-        let _ = self.stream.pause();
-    }
-}
