+++ audio/src/lib.rs
@@ -1,6 +1,3 @@
-use crate::audio::get_configuration;
-
-pub mod buffered_recorder;
 pub mod player;
 pub mod recorder;
 
@@ -13,9 +10,23 @@ mod pulse;
 mod cpal;
 
 pub fn get_audio_outputs() -> Vec<String> {
-    get_configuration().get_outputs()
+    #[cfg(target_os = "linux")]
+    {}
+
+    #[cfg(not(target_os = "linux"))]
+    {
+        use crate::cpal::cpal_config::CpalConfiguration;
+        CpalConfiguration::get_outputs()
+    }
 }
 
 pub fn get_audio_inputs() -> Vec<String> {
-    get_configuration().get_inputs()
+    #[cfg(target_os = "linux")]
+    {}
+
+    #[cfg(not(target_os = "linux"))]
+    {
+        use crate::cpal::cpal_config::CpalConfiguration;
+        CpalConfiguration::get_inputs()
+    }
 }
