+++ audio/src/player.rs
@@ -9,7 +9,7 @@ use std::path::PathBuf;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 
-use crate::audio::get_output;
+use crate::audio::{get_output, AudioSpecification};
 use symphonia::core::audio::SampleBuffer;
 use symphonia::core::errors::Error;
 use symphonia::core::formats::{SeekMode, SeekTo};
@@ -226,12 +226,18 @@ impl Player {
                     // Is this the first decoded packet?
                     if audio_output.is_none() && sample_buffer.is_none() {
                         let spec = *decoded.spec();
-                        let duration = decoded.capacity() as u64;
+                        let capacity = decoded.capacity() as u64;
 
-                        sample_buffer = Some(SampleBuffer::<f32>::new(duration, spec));
+                        sample_buffer = Some(SampleBuffer::<f32>::new(capacity, spec));
 
                         if !self.process_only {
-                            audio_output.replace(get_output(spec, self.device.clone())?);
+                            let audio_spec = AudioSpecification {
+                                device: self.device.clone(),
+                                spec,
+                                buffer: capacity as usize,
+                            };
+
+                            audio_output.replace(get_output(audio_spec)?);
                         }
                     }
 
@@ -241,9 +247,7 @@ impl Player {
                         let mut samples = buf.samples().to_vec();
 
                         if let Some(ref mut ebu_r128) = ebu_r128 {
-                            // We need to convert the samples to f64..
-                            let samples64: Vec<f64> = samples.iter().map(|n| *n as f64).collect();
-                            ebu_r128.add_frames_f64(samples64.as_slice())?;
+                            ebu_r128.add_frames_f32(samples.as_slice())?;
 
                             // Skip straight to the next packet..
                             continue;
